third_party/blink/web_tests/platform/generic/virtual/webrtc-wpt-plan-b/external/wpt/webrtc/RTCConfiguration-iceTransportPolicy-expected.txt

This is a testharness.js-based test.
PASS new RTCPeerConnection() should have default iceTransportPolicy all
PASS new RTCPeerConnection({ iceTransportPolicy: undefined }) should have default iceTransportPolicy all
PASS new RTCPeerConnection({ iceTransportPolicy: 'all' }) should succeed
PASS new RTCPeerConnection({ iceTransportPolicy: 'relay' }) should succeed
PASS setConfiguration({ iceTransportPolicy: 'relay' }) with initial iceTransportPolicy all should succeed
PASS setConfiguration({ iceTransportPolicy: 'all' }) with initial iceTransportPolicy relay should succeed
PASS setConfiguration({}) with initial iceTransportPolicy relay should set new value to all
PASS new RTCPeerConnection(config) - with invalid iceTransportPolicy should throw TypeError
PASS setConfiguration(config) - with invalid iceTransportPolicy should throw TypeError
PASS new RTCPeerConnection(config) - with none iceTransportPolicy should throw TypeError
PASS setConfiguration(config) - with none iceTransportPolicy should throw TypeError
PASS new RTCPeerConnection(config) - with null iceTransportPolicy should throw TypeError
PASS setConfiguration(config) - with null iceTransportPolicy should throw TypeError
FAIL new RTCPeerConnection({ iceTransports: 'relay' }) should have no effect assert_equals: expected "all" but got "relay"
FAIL new RTCPeerConnection({ iceTransports: 'invalid' }) should have no effect Failed to construct 'RTCPeerConnection': Failed to read the 'iceTransports' property from 'RTCConfiguration': The provided value 'invalid' is not a valid enum value of type RTCIceTransportPolicy.
FAIL new RTCPeerConnection({ iceTransports: null }) should have no effect Failed to construct 'RTCPeerConnection': Failed to read the 'iceTransports' property from 'RTCConfiguration': The provided value 'null' is not a valid enum value of type RTCIceTransportPolicy.
FAIL iceTransportPolicy "relay" on offerer should prevent candidate gathering promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
FAIL iceTransportPolicy "relay" on answerer should prevent candidate gathering promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
FAIL Changing iceTransportPolicy from "all" to "relay" causes an ICE restart which should fail, with no new candidates promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
FAIL Changing iceTransportPolicy from "relay" to "all" causes an ICE restart which should succeed promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
FAIL Changing iceTransportPolicy from "all" to "relay", and back to "all" prompts an ICE restart promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
FAIL Changing iceTransportPolicy from "all" to "relay" on the answerer has no effect on a subsequent offer/answer promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
Harness: the test ran to completion.
